useAdMob

+++ src/screens/Home/BookDetail/BookDetailContainer.tsx
@@ -7,6 +7,7 @@ import {
 } from '@src/navigation/HomeNavigator/route';
 import { useBookDetail } from './useBookDetail';
 import { useToast } from '@src/hooks';
+import { Alert } from 'react-native';
 
 export const BookDetailContainer: FC = () => {
   const navigation = useHomeNavigation();
@@ -17,16 +18,27 @@ export const BookDetailContainer: FC = () => {
   const { deleteBook, loadingDelete } = useBookDetail(route.params);
 
   const onDeleteBook = useCallback(async () => {
-    try {
-      await deleteBook();
-      showToast({
-        message: `${bookInfo.title}を削除しました`,
-        status: 'success',
-      });
-      navigation.dispatch(StackActions.popToTop());
-    } catch {
-      showToast({ message: `エラーが起きました`, status: 'error' });
-    }
+    Alert.alert('本当に削除しますか？', '', [
+      {
+        text: '削除',
+        onPress: async () => {
+          try {
+            await deleteBook();
+            showToast({
+              message: `${bookInfo.title}を削除しました`,
+              status: 'success',
+            });
+            navigation.dispatch(StackActions.popToTop());
+          } catch {
+            showToast({ message: `エラーが起きました`, status: 'error' });
+          }
+        },
+      },
+      {
+        text: 'キャンセル',
+        style: 'cancel',
+      },
+    ]);
   }, [showToast, deleteBook, navigation]);
 
   const back = useCallback(() => {
